Move helper methods to util file, piggyback for superscript, update readme.
